Support providing the db name to the subscriber? method
Previously, subscriber? provides strange results when you have the subscriptions
and publications in different databases on the same database cluster.

sub.subscriber? was true as expected
pub.subscriber? was also true and not expected

Now, we you can provide the db name to clarify which database you're asking if
they are a subscriber.

sub.subscriber?(sub.db) is now true as expected
pub.subscriber?(pub.db) is NOW false since the subscription is not created in
the pub.db.

The existing default of nil dbname is kept to avoid a backward incompatible
change.